Implement a global staff channel Fixes #15
Added a new global staff channel for staff members to communicate with all servers. When a message is received in the global staff channel, the bot will ignore any messages sent by bots and emit an event called 'globalStaffMessage'. In the 'index.js' file, a new listener has been added to handle the 'globalStaffMessage' event. When the event is triggered, the bot will send the message to all connected peers via WebSocket.